actualizo backv1

+++ backendQS/AppQuickSolutions/src/main/java/com/QS/AppQuickSolutions/services/PartService.java
@@ -1,5 +1,7 @@
 package com.QS.AppQuickSolutions.services;
 
+import java.io.IOException;
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.UUID;
 
@@ -8,12 +10,9 @@ import org.springframework.stereotype.Service;
 
 import com.QS.AppQuickSolutions.dto.PartDto;
 import com.QS.AppQuickSolutions.entity.Part;
-import com.QS.AppQuickSolutions.entity.User;
-import com.QS.AppQuickSolutions.enums.Material;
-import com.QS.AppQuickSolutions.enums.PartName;
-import com.QS.AppQuickSolutions.enums.Role;
+import com.QS.AppQuickSolutions.entity.Project;
 import com.QS.AppQuickSolutions.repository.PartRepository;
-import com.QS.AppQuickSolutions.repository.ProjectRepository;
+import com.google.zxing.WriterException;
 
 import jakarta.persistence.EntityNotFoundException;
 
@@ -24,46 +23,55 @@ public class PartService {
     private PartRepository partRepository;
 
     @Autowired
-    private ProjectRepository projectRepository;
+    private QRCodeService qrCodeService;
 
-    public Part updatePart(UUID id, PartDto partDto, User user) {
-        Part existingPart = partRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Part not found"));
+    public Part createPart(Project project, PartDto partDto) throws IOException, WriterException {
+        Part part = new Part();
 
-        if (user.getRole() == Role.ADMIN) {
-            existingPart.setPartName(PartName.valueOf(partDto.getPartName()));
-            existingPart.setMaterial(Material.valueOf(partDto.getMaterial()));
-            existingPart.setTotalweightKg(partDto.getTotalweightKg());
-            existingPart.setSheetThicknessMm(partDto.getSheetThicknessMm());
-            existingPart.setLengthPiecesMm(partDto.getLengthPiecesMm());
-            existingPart.setHeightMm(partDto.getHeightMm());
-            existingPart.setWidthMm(partDto.getWidthMm());
-        }
+        part.setProject(project);
+        part.setCustomPart(partDto.getCustomPart());
+        part.setPartMaterial(partDto.getPartMaterial());
+        part.setTotalweightKg(partDto.getTotalweightKg());
+        part.setSheetThicknessMm(partDto.getSheetThicknessMm());
+        part.setLengthPiecesMm(partDto.getLengthPiecesMm());
+        part.setHeightMm(partDto.getHeightMm());
+        part.setWidthMm(partDto.getWidthMm());
+        part.setObservations(partDto.getObservations());
 
-        if (user.getRole() == Role.ADMIN || user.getRole() == Role.OPERATOR) {
-            existingPart.setScanDateTime(partDto.getScanDateTime());
-            existingPart.setReceptionState(partDto.getReceptionState());
-            existingPart.setQualityControlState(partDto.getQualityControlState());
-            existingPart.setObservations(partDto.getObservations());
-        }
+        // Generar datos para el QR
+        String qrData = "Part ID: " + UUID.randomUUID() +
+                "\nProject: " + project.getProjectName() +
+                "\nCustomPart: " + partDto.getCustomPart().getCustomPart() +
+                "\nMaterial: " + partDto.getPartMaterial().getMaterialName();
+        part.setQrCodeData(qrData);
 
-        return partRepository.save(existingPart);
+        // Generar QR
+        String qrFilePath = qrCodeService.generateQRCodeImage(qrData, 300, 300, UUID.randomUUID() + "_part_qr.png");
+        part.setQrCodeFilePath(qrFilePath);
+
+        return partRepository.save(part);
     }
 
     public List<Part> getAllParts() {
         return partRepository.findAll();
     }
 
-     public List<Part> getPartsByProject(Long projectId) {
-        // Verifica si el proyecto existe
-        if (!projectRepository.existsById(projectId)) {
-            throw new EntityNotFoundException("Proyecto no encontrado con ID: " + projectId);
-        }
+    public Part getPartById(UUID id) {
+        return partRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("Part no encontrada con ID: " + id));
+    }
 
-        // Obtiene las partes asociadas al proyecto
-        return partRepository.findByProjectId(projectId);
+    public Part scanPart(UUID id) {
+        Part part = getPartById(id);
+        part.setReceptionState(true);
+        part.setScanDateTime(LocalDateTime.now());
+        return partRepository.save(part);
     }
 
     public void deletePart(UUID id) {
+        if (!partRepository.existsById(id)) {
+            throw new EntityNotFoundException("Part no encontrada con ID: " + id);
+        }
         partRepository.deleteById(id);
     }
 }
